test(claude-sdk-oauth): share the #6981 restart fixture between the lifecycle and compaction regressions
Both regression files carried the same session/extension fixture; hoisting it
into test/helpers/claude-sdk-oauth-restart-fixture.ts keeps each file under the
250 pure-LOC ceiling and lets the two suites drift together.
